new synclk config + updated results

+++ run/baseline_eval.sh
@@ -6,15 +6,22 @@ set -x
 circuit=${1?"Specify a circuit as the first positional argument"}
 library=${2?"Specify a library as the second positional argument. Options are: asap7, variability14, nangate45"}
 
-export synclk="0.0"
-incr="0.01"
-export top_design="top"
-
 maindir="$HOME/axcarbon"
 testdir="$maindir"
 cd $testdir
-
 circdir="$maindir/circuits/$circuit"
+top_design="top"
+
+# functions to get area, delay and power
+function getArea {
+    awk '/Total cell area/ {print $NF}' $testdir/reports/${top_design}_${synclk}${tunit}.area.rpt
+}
+function getDelay {
+    grep "data arrival time" $testdir/reports/${top_design}_${synclk}${tunit}.timing.pt.rpt | awk 'NR==1 {print $NF}'
+}
+function getPower {
+    awk '/Total Power/ {print $4}' $testdir/reports/${top_design}_${synclk}${tunit}.power.ptpx.rpt
+}
 
 # set up libraries and environment
 if [[ $library == "asap7" ]]; then
@@ -56,7 +63,6 @@ export tunit="$tunit"
 sed -i "/ENV_LIBRARY_PATH=/ c\export ENV_LIBRARY_PATH=\"$libpath\"" ./scripts/env.sh
 sed -i "/ENV_LIBRARY_DB=/ c\export ENV_LIBRARY_DB=\"$lib\"" ./scripts/env.sh
 sed -i "/ENV_LIBRARY_VERILOG_PATH=/ c\export ENV_LIBRARY_VERILOG_PATH=\"$libverilog\"" ./scripts/env.sh
-sed -i "/ENV_CLK_PERIOD=/ c\export ENV_CLK_PERIOD=\"$synclk\"" ./scripts/env.sh
 
 # prepare C libraries
 if [ ! -f "$libcpath/library.c" ] || [ ! -f "$libcpath/library.h" ]; then
@@ -68,10 +74,11 @@ rm -f $maindir/libs/*.so
 cp $libcpath/library.c $maindir/libs/
 cp $libcpath/library.h $maindir/libs/
 
+# prepare the results directory and file
 mkdir -p $testdir/results
 mkdir -p $testdir/results/baseline
 resfile=$testdir/results/baseline/${circuit}.txt
-echo -e "Area\tDelay\tError\tPower" > $resfile
+echo -e "SynClk\tArea\tDelay\tError\tPower" > $resfile
 
 # prepare testbench, inputs and true outputs for simulation
 cp $circdir/top.v ./hdl/top.v
@@ -79,43 +86,48 @@ cp $circdir/tb.v ./sim/top_tb.v
 awk -F'_' '{for (i=1; i<NF; i++) {printf("%s", $i); if(i<NF-1) printf("_")} printf("\n")}' $circdir/inputs.txt > ./sim/inputs.txt
 awk -F'_' '{print $NF}' $circdir/inputs.txt > ./sim/expected.txt
 
-# reports
-area_rpt="$testdir/reports/${top_design}_${synclk}${tunit}.area.rpt"
-delay_rpt="$testdir/reports/${top_design}_${synclk}${tunit}.timing.pt.rpt"
-power_rpt="$testdir/reports/${top_design}_${synclk}${tunit}.power.ptpx.rpt"
-
-# get area from synthesis
-make dcsyn
-area=$(awk '/Total cell area/ {print $NF}' $area_rpt)
-
-# sta to get delay
-make sta
-delay="$(grep "data arrival time" $delay_rpt | awk 'NR==1 {print $NF}')"
-
-# Iteratively, perform gate level simulations to find the simulation clock which does not produce
-# any timing errors. That would be the circuit's functional Tclk
-# NOTE: 'true' may not be portable, so we use integer comparison
+synclk="0.0"
+delay_incr="0.01"
 while true; do
-   sed -i "/parameter PERIOD=/ c\parameter PERIOD=$delay;" ./sim/top_tb.v
 
-   rm -rf work_gate_lib
-   rm -rf gate_simv.daidir
-   rm -rf tech_lib/
+    # Iteratively, perform synthesis and STA to find the minimum clock period that the circuit can match
+    sed -i "/ENV_CLK_PERIOD=/ c\export ENV_CLK_PERIOD=\"$synclk\"" ./scripts/env.sh
+    make dcsyn
+    area="$(getArea)"
+    make sta
+    delay="$(getDelay)"
 
-   make gate_sim
-   error="$(./scripts/errors/med.sh ./sim/expected.txt ./sim/output.txt)"
-   if [ 1 -eq "$(awk -v e=$error 'BEGIN {if (e>0) print "1"; else print "0"}')" ]; then
-       delay="$(awk -v d=$delay -v incr=$incr 'BEGIN { print d + incr }')"
-   else
-       break
-   fi
+    if [ 1 -eq "$(awk -v d=$delay -v s=$synclk 'BEGIN {if (d>s) print "1"; else print "0"}')" ]; then
+        synclk="$(echo $delay | awk '{printf "%.2f\n", $1}')"
+        continue
+    fi
+
+    # Iteratively, perform gate level simulations to find the simulation clock which does not produce
+    # any timing errors. That would be the circuit's functional Tclk
+    # NOTE: 'true' may not be portable, so we use integer comparison
+
+    sed -i "/parameter PERIOD=/ c\parameter PERIOD=$delay;" ./sim/top_tb.v
+    rm -rf work_gate_lib
+    rm -rf gate_simv.daidir
+    rm -rf tech_lib/
+
+    make gate_sim
+    error="$(./scripts/errors/med.sh ./sim/expected.txt ./sim/output.txt)"
+
+    if [ 1 -eq "$(awk -v e=$error 'BEGIN {if (e>0) print "1"; else print "0"}')" ]; then
+        synclk="$(awk -v d=$delay -v incr=$delay_incr 'BEGIN {print d+incr}' | awk '{printf "%.2f\n", $1}')"
+    else
+        # break only if the error is zero and the delay is less than the synthesis delay (checked above)
+        break
+    fi
+
 done
 # error="0.0"
 
 # get power
 make power
-power=$(awk '/Total Power/ {print $4}' $power_rpt)
+power="$(getPower)"
 # power="0.0"
 
-echo -e "$area\t$delay\t$error\t$power" >> $resfile
+echo -e "$synclk\t$area\t$delay\t$error\t$power" >> $resfile
